showkey: Fix build warning
showkey.c: In function ‘main’:
showkey.c:275:31: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  275 |                         alarm(timeout);
      |                               ^~~~~~~
showkey.c:287:23: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  287 |                 alarm(timeout);
      |                       ^~~~~~~

+++ src/showkey.c
@@ -272,7 +272,7 @@ int main(int argc, char *argv[])
 	/* show scancodes */
 	if (!show_keycodes) {
 		while (1) {
-			alarm(timeout);
+			alarm((unsigned int) timeout);
 			n = read(fd, buf, sizeof(buf));
 			for (i = 0; i < n; i++)
 				printf("0x%02x ", buf[i]);
@@ -284,7 +284,7 @@ int main(int argc, char *argv[])
 
 	/* show keycodes - 2.6 allows 3-byte reports */
 	while (1) {
-		alarm(timeout);
+		alarm((unsigned int) timeout);
 		n = read(fd, buf, sizeof(buf));
 		i = 0;
 		while (i < n) {
